Move duplicatelabels.ts to src/components/ because it does not implement abstract providers defined in VS Code API.
