+++ src/main.ts
@@ -25,7 +25,7 @@ import {UtensilsParser as PEGParser} from './components/parser/syntax'
 import {Completer} from './providers/completion'
 import {BibtexCompleter} from './providers/bibtexcompletion'
 import {CodeActions} from './providers/codeactions'
-import {DuplicateLabels} from './providers/duplicatelabels'
+import {DuplicateLabels} from './components/duplicatelabels'
 import {HoverProvider} from './providers/hover'
 import {GraphicsPreview} from './providers/preview/graphicspreview'
 import {MathPreview} from './providers/preview/mathpreview'
